mysite/requirements.txt
antiorm==1.2.1
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==43.0.3
db==0.1.1
db-sqlite3==0.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.2
django-allauth==65.2.0
django-extensions==3.2.3
docker==7.1.0
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.21.1
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaleido==0.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mysqlclient==2.2.5
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdfkit==1.0.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rpds-py==0.22.0
scipy==1.14.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tenacity==9.0.0
text-unidecode==1.3
tinycss2==1.4.0
tk==0.1.0
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
